cloud-run/requirements.txt
flask==3.1.0
google-api-python-client==2.163.0
google-auth==2.38.0
google-cloud-bigquery==3.27.0
pandas==2.2.3
pyarrow==18.1.0
gunicorn==23.0.0
